Now only creating integral sized images. Now returning an image with a bitmap image rep instead of a cached rep.

+++ MacOSaiX/NSImage+MacOSaiX.h
@@ -11,6 +11,6 @@
 
 @interface NSImage (MacOSaiX)
 
-- (NSImage *)copyWithLargestDimension:(float)size;
+- (NSImage *)copyWithLargestDimension:(int)size;
 
 @end
